backend/chat_langchain_service/requirements.txt
# Core FastAPI and async support
fastapi==0.115.12
uvicorn[standard]==0.34.1
starlette==0.46.2

# Database and ORM
sqlmodel==0.0.24
SQLAlchemy==2.0.40
alembic==1.12.1
psycopg2-binary==2.9.10

# Redis for caching
redis==5.2.1

# Message Queue
pika==1.3.2

# HTTP clients with async support
httpx==0.28.1
aiohttp==3.10.11
requests==2.32.3

# Authentication and security
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
cryptography==44.0.2

# LangChain and AI
langchain-core>=0.3.17,<0.4.0
langchain>=0.3.7
langchain-openai>=0.2.8
langchain-community==0.3.5
openai==1.54.4

# Kubernetes client
kubernetes==32.0.1

# Data validation and serialization
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1

# Configuration and environment
python-dotenv==1.1.0
PyYAML==6.0.2

# Utilities and parsing
python-multipart==0.0.20
email_validator==2.2.0
python-dateutil==2.9.0.post0

# Production monitoring and logging
structlog==24.4.0
prometheus-client==0.21.0

# Performance and reliability
tenacity==9.0.0
backoff==2.2.1
circuit-breaker==0.1.1

# Security and validation
validators==0.34.0
bleach==6.2.0

# Async utilities
asyncio-throttle==1.0.2
aiofiles==24.1.0

# System dependencies
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
urllib3==2.4.0
six==1.17.0
setuptools==79.0.0

# XML and HTML processing
lxml==5.3.2
cssselect==1.3.0
pyquery==2.0.1

# Async and concurrency
gevent==25.4.1
greenlet==3.2.0
anyio==3.7.1
sniffio==1.3.1

# Cryptography and encoding
ecdsa==0.19.1
rsa==4.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.22

# Google services
google-auth==2.39.0
cachetools==5.5.2
oauthlib==3.2.2
requests-oauthlib==2.0.0

# Template and markup
Mako==1.3.10
MarkupSafe==3.0.2

# Networking and protocols
dnspython==2.7.0
websocket-client==1.8.0
h11==0.14.0
httpcore==1.0.8

# Utilities
click==8.1.8
tqdm==4.67.1
distro==1.9.0
durationpy==0.9
proxies==1.6
typing-inspection==0.4.0
typing_extensions==4.13.2

# Zope utilities
zope.event==5.0
zope.interface==7.2

# Configuration utilities
ruamel.appconfig==0.5.5
ruamel.std.argparse==0.8.3

# Additional annotations
annotated-types==0.7.0
